chore(components): rename modal, timerange, prompt
