+++ packages/components/src/index.ts
@@ -30,8 +30,8 @@ export * from './components/Warning/Warning';
 export * from './components/others/Box';
 export * from './components/others/Card';
 export * from './components/others/SecurityCard';
-export * from './components/prompts/ConfirmOnDevice';
-export * from './components/Modal';
+export * from './components/prompts/ConfirmOnDevice/ConfirmOnDevice';
+export * from './components/Modal/Modal';
 export * from './components/Modal/Backdrop';
 export * from './components/logos/CoinLogo/CoinLogo';
 export * from './components/logos/TrezorLogo/TrezorLogo';
@@ -40,7 +40,7 @@ export * from './components/loaders/Progress';
 export * from './components/loaders/FluidSpinner';
 export * from './components/loaders/LoadingContent';
 export * from './components/Tooltip';
-export * from './components/Timerange';
+export * from './components/Timerange/Timerange';
 export * from './components/Truncate';
 export * from './components/form/SelectBar';
 export * from './components/HoverAnimation';
